+++ src/components/SQForm/SQFormDropdown.js
@@ -138,7 +138,11 @@ SQFormDropdown.propTypes = {
   children: PropTypes.arrayOf(
     PropTypes.shape({
       label: PropTypes.string,
-      value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+      value: PropTypes.oneOfType([
+        PropTypes.string,
+        PropTypes.number,
+        PropTypes.bool
+      ]),
       isDisabled: PropTypes.bool
     })
   ),