+++ src/components/SQForm/SQFormReadOnlyField.tsx
@@ -22,7 +22,7 @@ function SQFormReadOnlyField({
   return (
     <Grid item sm={size}>
       <TextField
-        id={toKebabCase(label)}
+        id={toKebabCase(name)}
         label={label}
         name={name}
         value={field.value || '- -'}