+++ stories/SQFormGuidedWorkflow.stories.js
@@ -1,0 +1,169 @@
+import React from 'react';
+import * as Yup from 'yup';
+import {
+  ExpandingCardList,
+  ExpandingCard,
+  RoundedButton,
+  TextButton
+} from 'scplus-shared-components';
+import {SQFormGuidedWorkflow, SQFormDropdown, SQFormTextarea} from '../src';
+
+const sleep = milliseconds => {
+  return new Promise(resolve => setTimeout(resolve, milliseconds));
+};
+
+export default {
+  title: 'Forms/SQFormGuidedWorkflow',
+  component: SQFormGuidedWorkflow
+};
+
+const outcomeDropdownOptions = [
+  {label: 'Not Interested', value: 'not-interested'},
+  {label: 'Interested', value: 'interested'}
+];
+
+const Template = () => {
+  const [isIneligible, setIneligible] = React.useState(false);
+  const scriptedTextMap = {
+    customerName: 'Bob Smith',
+    agentName: 'Jane Doe',
+    planName: 'Super Cheap Med+'
+  };
+  const taskModules = [
+    {
+      name: 'intro',
+      title: 'Introduction',
+      formikProps: {
+        initialValues: {
+          outcome: '',
+          notes: ''
+        },
+        onSubmit: async values => {
+          await sleep(3000); // Simulate API call to see loading spinner
+          console.log(JSON.stringify(values));
+        },
+        validationSchema: {
+          outcome: Yup.string().required('Required'),
+          notes: Yup.string()
+        }
+      },
+      scriptedTextProps: {
+        text: `Hi, ${scriptedTextMap.customerName}, my name is ${scriptedTextMap.agentName}, and I am calling to discuss ${scriptedTextMap.planName}.\n Are you available right now to talk through some things with me, today?`,
+        title: 'Agent Script'
+      },
+      outcomeProps: {
+        FormElements: (
+          <>
+            <SQFormDropdown name="outcome" label="Outcome" isRequired={true}>
+              {outcomeDropdownOptions}
+            </SQFormDropdown>
+            <SQFormTextarea name="notes" label="Notes" />
+          </>
+        ),
+        title: 'Confirm Info'
+      }
+    },
+    {
+      name: 'cancellation',
+      title: 'Policy Cancellation Statement',
+      subtitles: ['Cancelling policy'],
+      isFailedState: isIneligible,
+      formikProps: {
+        initialValues: {
+          outcome: '',
+          notes: ''
+        },
+        onSubmit: async values => {
+          console.log(JSON.stringify(values));
+        },
+        validationSchema: {
+          outcome: Yup.string().required('Required'),
+          notes: Yup.string()
+        }
+      },
+      scriptedTextProps: {
+        text: `Stuff about policy cancellation documents`,
+        title: 'Agent Script',
+        actionButton: <TextButton tooltip="View">View Doc</TextButton>
+      },
+      outcomeProps: {
+        FormElements: (
+          <>
+            <SQFormDropdown name="outcome" label="Outcome" isRequired={true}>
+              {outcomeDropdownOptions}
+            </SQFormDropdown>
+            <SQFormTextarea name="notes" label="Notes" />
+            <RoundedButton
+              title="Ineligible"
+              onClick={() => setIneligible(true)}
+            >
+              Ineligible? Click here
+            </RoundedButton>
+          </>
+        ),
+        title: 'Confirm Info',
+        infoText:
+          'Interact with the form to see me change colors based on form state',
+        warningText: 'Form needs your attention',
+        successText: 'Form is ready to submit',
+        errorText: 'Do not pass go, do not collect $200'
+      }
+    },
+    {
+      name: 'providers',
+      title: 'Verify Providers',
+      subtitles: ['Please verify the providers'],
+      formikProps: {
+        initialValues: {
+          outcome: '',
+          notes: ''
+        },
+        onSubmit: async values => {
+          console.log(JSON.stringify(values));
+        },
+        validationSchema: {
+          outcome: Yup.string().required('Required'),
+          notes: Yup.string()
+        }
+      },
+      scriptedTextProps: {
+        text:
+          'Before proceeding, please verify the list of providers is correct',
+        title: 'Agent Script',
+        actionButton: <TextButton tooltip="View">View Providers</TextButton>
+      },
+      outcomeProps: {
+        FormElements: (
+          <>
+            <SQFormDropdown name="outcome" label="Outcome" isRequired={true}>
+              {outcomeDropdownOptions}
+            </SQFormDropdown>
+            <SQFormTextarea name="notes" label="Notes" />
+          </>
+        ),
+        title: 'Confirm Info'
+      }
+    }
+  ];
+
+  return (
+    <div style={{width: '90%', height: '95vh'}}>
+      <ExpandingCardList>
+        <ExpandingCard title="Guided Workflow" name="guidedWorkflow">
+          <SQFormGuidedWorkflow
+            mainTitle="CCA Guided Workflow"
+            mainSubtitle="Please review these Services with you client, then confirm their responses."
+            initialCompletedTasks={0}
+            isStrictMode={false}
+            taskModules={taskModules}
+            onError={error => {
+              console.error(error);
+            }}
+          />
+        </ExpandingCard>
+      </ExpandingCardList>
+    </div>
+  );
+};
+
+export const Default = Template.bind({});