+++ stories/__tests__/SQFormButton.stories.test.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import {render, screen} from '@testing-library/react';
+import {fireEvent, render, screen} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import {composeStories} from '@storybook/testing-react';
 import * as stories from '../SQFormButton.stories';
@@ -94,7 +94,7 @@ describe('SQFormButton Tests', () => {
         name: /form submission/i
       });
 
-      userEvent.click(submitButton);
+      fireEvent.click(submitButton);
 
       expect(onClickSpy).not.toHaveBeenCalled();
     });