+++ CHANGELOG.md
@@ -1,3 +1,98 @@
+### [5.2.2](https://github.com/SelectQuoteLabs/SQForm/compare/v5.2.1...v5.2.2) (2021-08-02)
+
+
+### Bug Fixes
+
+* 🐛 gave autocomplete dropdown tooltips and larger size ([13dff61](https://github.com/SelectQuoteLabs/SQForm/commit/13dff615b4849b6134a0e32ec080a5d8e22768e0))
+
+### [5.2.1](https://github.com/SelectQuoteLabs/SQForm/compare/v5.2.0...v5.2.1) (2021-08-02)
+
+
+### Bug Fixes
+
+* **SQFormGuidedWorkflow.js:** add padding and update shared component version ([3eebe22](https://github.com/SelectQuoteLabs/SQForm/commit/3eebe2218c667bac12e729a03c3153f2830c5769))
+
+## [5.2.0](https://github.com/SelectQuoteLabs/SQForm/compare/v5.1.0...v5.2.0) (2021-07-20)
+
+
+### Features
+
+* 🎸 Guided Workflow feature ([777346a](https://github.com/SelectQuoteLabs/SQForm/commit/777346acd40718ef5258f127cf51e0c6d3f349cd))
+
+## [5.1.0](https://github.com/SelectQuoteLabs/SQForm/compare/v5.0.3...v5.1.0) (2021-07-19)
+
+
+### Features
+
+* 🎸 add `stripNonNumeric` prop to `SQFormMaskedTextField` ([c770072](https://github.com/SelectQuoteLabs/SQForm/commit/c7700720cd245d976fc118939ba024f3354398c3)), closes [#224](https://github.com/SelectQuoteLabs/SQForm/issues/224)
+
+### [5.0.3](https://github.com/SelectQuoteLabs/SQForm/compare/v5.0.2...v5.0.3) (2021-07-15)
+
+
+### Bug Fixes
+
+* 🐛 correctly pass props to Icon ([01df7fa](https://github.com/SelectQuoteLabs/SQForm/commit/01df7fa4526270bb96661da04e4d2ea74bb6bff2)), closes [#331](https://github.com/SelectQuoteLabs/SQForm/issues/331)
+
+### [5.0.2](https://github.com/SelectQuoteLabs/SQForm/compare/v5.0.1...v5.0.2) (2021-06-30)
+
+
+### Bug Fixes
+
+* 🐛 Removed unused props from the stepper component ([ad6df58](https://github.com/SelectQuoteLabs/SQForm/commit/ad6df585ad60a36439e73393c70abe23b75b47b1)), closes [#279](https://github.com/SelectQuoteLabs/SQForm/issues/279)
+
+### [5.0.1](https://github.com/SelectQuoteLabs/SQForm/compare/v5.0.0...v5.0.1) (2021-06-25)
+
+
+### Bug Fixes
+
+* 🐛 Updated id values to name instead of label for more uniq ([14764b9](https://github.com/SelectQuoteLabs/SQForm/commit/14764b9520fd8bd7d0e4c5c7885e8b4b223ec656)), closes [#318](https://github.com/SelectQuoteLabs/SQForm/issues/318)
+
+## [5.0.0](https://github.com/SelectQuoteLabs/SQForm/compare/v4.9.0...v5.0.0) (2021-06-22)
+
+
+### ⚠ BREAKING CHANGES
+
+* 🧨 Adding 20px padding to SQFormDialog content instead of 8px 24px
+
+### Features
+
+* 🎸 Adding 20px padding to dialog content ([24df034](https://github.com/SelectQuoteLabs/SQForm/commit/24df0347d1870a44048ca049177ec77924662ba8)), closes [#88](https://github.com/SelectQuoteLabs/SQForm/issues/88)
+
+## [4.9.0](https://github.com/SelectQuoteLabs/SQForm/compare/v4.8.1...v4.9.0) (2021-06-22)
+
+
+### Features
+
+* 🎸 support for boolean valued dropdown ([cc2dca9](https://github.com/SelectQuoteLabs/SQForm/commit/cc2dca939b0ad4ecbd783154738bbc45a7c5e2b5)), closes [#174](https://github.com/SelectQuoteLabs/SQForm/issues/174)
+
+### [4.8.1](https://github.com/SelectQuoteLabs/SQForm/compare/v4.8.0...v4.8.1) (2021-06-18)
+
+
+### Bug Fixes
+
+* 🐛 SQFormMultiValue accepts custom initial values ([632ed3c](https://github.com/SelectQuoteLabs/SQForm/commit/632ed3c007d9aff9dad8c1c0dde23ebdd2ee22fb)), closes [#308](https://github.com/SelectQuoteLabs/SQForm/issues/308)
+
+## [4.8.0](https://github.com/SelectQuoteLabs/SQForm/compare/v4.7.0...v4.8.0) (2021-06-17)
+
+
+### Features
+
+* 🎸 Add SQFormMultiValue component ([50583c9](https://github.com/SelectQuoteLabs/SQForm/commit/50583c9887326c2945d841cd59b5dcbfc982e8a8)), closes [#303](https://github.com/SelectQuoteLabs/SQForm/issues/303) [#304](https://github.com/SelectQuoteLabs/SQForm/issues/304)
+
+## [4.7.0](https://github.com/SelectQuoteLabs/SQForm/compare/v4.6.0...v4.7.0) (2021-06-04)
+
+
+### Features
+
+* 🎸 ScrollableCard that always maintains full height ([584aba2](https://github.com/SelectQuoteLabs/SQForm/commit/584aba23482b6254653143adb56e49e534613757)), closes [#226](https://github.com/SelectQuoteLabs/SQForm/issues/226)
+
+## [4.6.0](https://github.com/SelectQuoteLabs/SQForm/compare/v4.5.17...v4.6.0) (2021-06-02)
+
+
+### Features
+
+* 🎸 fix test for icon button ([0228bb8](https://github.com/SelectQuoteLabs/SQForm/commit/0228bb8cc1909a74f51775553e737a4560ced08e))
+
 ### [4.5.17](https://github.com/SelectQuoteLabs/SQForm/compare/v4.5.16...v4.5.17) (2021-05-21)